[Rename] Move Fragments to place_list directory
