+++ app/src/main/java/com/twoday/todaytrip/ui/place_list/FirstRecyclerViewFragment.kt
@@ -1,4 +1,4 @@
-package com.twoday.todaytrip.ui
+package com.twoday.todaytrip.ui.place_list
 
 import android.os.Bundle
 import android.view.LayoutInflater
