+++ app/src/main/java/com/twoday/todaytrip/ui/place_list/SecondRecyclerViewFragment.kt
@@ -1,0 +1,20 @@
+package com.twoday.todaytrip.ui.place_list
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.twoday.todaytrip.R
+
+class SecondRecyclerViewFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_second_recycler_view, container, false)
+    }
+
+}
