+++ app/src/main/java/com/twoday/todaytrip/ui/place_list/ThirdRecyclerViewFragment.kt
@@ -1,4 +1,4 @@
-package com.twoday.todaytrip.ui
+package com.twoday.todaytrip.ui.place_list
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.twoday.todaytrip.R
 
-class SecondRecyclerViewFragment : Fragment() {
+class ThirdRecyclerViewFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second_recycler_view, container, false)
+        return inflater.inflate(R.layout.fragment_third_recycler_view, container, false)
     }
 
 }
