+++ app/src/main/res/layout/fragment_event_recycler_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.EventRecyclerViewFragment">
+    tools:context=".ui.place_list.EventRecyclerViewFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
