+++ app/src/main/res/layout/fragment_first_recycler_view.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginHorizontal="16dp"
-    tools:context=".ui.FirstRecyclerViewFragment">
+    tools:context=".ui.place_list.FirstRecyclerViewFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_first_recycler_view"
