+++ app/src/main/res/layout/fragment_place_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.PlaceListFragment">
+    tools:context=".ui.place_list.PlaceListFragment">
 
     <TextView
         android:id="@+id/tv_travel_address"
