+++ app/src/main/res/layout/fragment_second_recycler_view.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.SecondRecyclerViewFragment">
+    tools:context=".ui.place_list.SecondRecyclerViewFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
